Get rid of compiler warnings on Visual Studio 2019.

+++ src/net.cpp
@@ -58,7 +58,7 @@ udp_server::~udp_server( )
 void udp_server::send( struct sockaddr_in &clientaddr_, const std::vector< uint8_t > &data_ )
 {
   int n =
-    sendto( m_fd, (const char *) data_.data(), data_.size(), 0, (struct sockaddr *) &clientaddr_, sizeof( clientaddr_ ) );
+    sendto( m_fd, (const char *) data_.data( ), (int) data_.size( ), 0, (struct sockaddr *) &clientaddr_, sizeof( clientaddr_ ) );
 
   if ( n != data_.size() )
   {
@@ -72,7 +72,7 @@ bool udp_server::receive( struct sockaddr_in &clientaddr_, std::vector< uint8_t 
   data_.resize( 16 * 1024 ); // TODO: Detect MTU
   socklen_t clientlen = sizeof( clientaddr_ );
 
-  int n = recvfrom( m_fd, (char *) data_.data(), data_.size(), 0, (struct sockaddr *) &clientaddr_, &clientlen );
+  int n = recvfrom( m_fd, (char *) data_.data( ), (int) data_.size( ), 0, (struct sockaddr *) &clientaddr_, &clientlen );
 
   if ( n < 0 )
   {
